Feature flag AC to make dev easier

+++ src/main/java/com/jftse/emulator/server/game/core/game/handler/GamePacketHandler.java
@@ -1,5 +1,6 @@
 package com.jftse.emulator.server.game.core.game.handler;
 
+import com.jftse.emulator.common.GlobalSettings;
 import com.jftse.emulator.common.utilities.BitKit;
 import com.jftse.emulator.common.utilities.StreamUtils;
 import com.jftse.emulator.common.utilities.StringUtils;
@@ -140,10 +141,13 @@ public class GamePacketHandler {
             account.setStatus((int) S2CLoginAnswerPacket.SUCCESS);
             authenticationService.updateAccount(account);
         });
-        List<ClientWhitelist> clientWhiteList = clientWhitelistService.findAll();
-        for (int i = 0; i < clientWhiteList.size(); i++) {
-            Long id = clientWhiteList.get(i).getId();
-            clientWhitelistService.remove(id);
+
+        if (GlobalSettings.IsAntiCheatEnabled) {
+            List<ClientWhitelist> clientWhiteList = clientWhitelistService.findAll();
+            for (int i = 0; i < clientWhiteList.size(); i++) {
+                Long id = clientWhiteList.get(i).getId();
+                clientWhitelistService.remove(id);
+            }
         }
 
         this.getGameHandler().getRoomList().clear();
@@ -1716,8 +1720,7 @@ public class GamePacketHandler {
     }
 
     public void handleDevPacket(Connection connection, Packet packet) {
-        boolean active = false; // Never change this to true for deployed versions
-        if (active) {
+        if (GlobalSettings.ShouldHandleDevPackets) {
             byte[] data = packet.getData();
             Packet packetToRelay = new Packet(data);
             this.getGameHandler().getClientList().forEach(x -> x.getConnection().sendTCP(packetToRelay));
@@ -1954,6 +1957,7 @@ public class GamePacketHandler {
     }
 
     public void handleHeartBeatPacket(Connection connection, Packet packet) {
+        if (!GlobalSettings.IsAntiCheatEnabled) return;
         String hostAddress = connection.getClient().getIp();
         ClientWhitelist clientWhitelist = clientWhitelistService.findByIpAndHwid(hostAddress, connection.getHwid());
         if (clientWhitelist == null)
